Use uint32_t for line numbers.
Hardly any extra memory is necessary for this, and it might be useful at some
point in the future.

+++ src/lily_emitter.h
@@ -121,7 +121,7 @@ typedef struct {
     uint32_t expr_num;
 
     /* This is the current line number, at any given time. */
-    uint16_t *lex_linenum;
+    uint32_t *lex_linenum;
 
     lily_raiser *raiser;
